@@ -1,6 +1,6 @@
 package io.github.jamsesso.jsonlogic;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -8,62 +8,62 @@ import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 
 public class InExpressionTests {
-  private static final JsonLogic jsonLogic = new JsonLogic();
+    private static final JsonLogic jsonLogic = new JsonLogic();
 
-  @Test
-  public void testStringIn() throws JsonLogicException {
-    assertEquals(true, jsonLogic.apply("{\"in\": [\"race\", \"racecar\"]}", null));
-  }
+    @Test
+    public void testStringIn() throws JsonLogicException {
+        assertEquals(true, jsonLogic.apply("{\"in\": [\"race\", \"racecar\"]}", null));
+    }
 
-  @Test
-  public void testStringNotIn() throws JsonLogicException {
-    assertEquals(false, jsonLogic.apply("{\"in\": [\"race\", \"clouds\"]}", null));
-  }
+    @Test
+    public void testStringNotIn() throws JsonLogicException {
+        assertEquals(false, jsonLogic.apply("{\"in\": [\"race\", \"clouds\"]}", null));
+    }
 
-  @Test
-  public void testArrayIn() throws JsonLogicException {
-    assertEquals(true, jsonLogic.apply("{\"in\": [1, [1, 2, 3]]}", null));
-    assertEquals(true, jsonLogic.apply("{\"in\": [4.56, [1, 2, 3, 4.56]]}", null));
-  }
+    @Test
+    public void testArrayIn() throws JsonLogicException {
+        assertEquals(true, jsonLogic.apply("{\"in\": [1, [1, 2, 3]]}", null));
+        assertEquals(true, jsonLogic.apply("{\"in\": [4.56, [1, 2, 3, 4.56]]}", null));
+    }
 
-  @Test
-  public void testArrayNotIn() throws JsonLogicException {
-    assertEquals(false, jsonLogic.apply("{\"in\": [5, [1, 2, 3]]}", null));
-  }
+    @Test
+    public void testArrayNotIn() throws JsonLogicException {
+        assertEquals(false, jsonLogic.apply("{\"in\": [5, [1, 2, 3]]}", null));
+    }
 
-  @Test
-  public void testInVariableInt() throws JsonLogicException {
-    Map data = Collections.singletonMap("list", Arrays.asList(1, 2, 3));
-    assertEquals(true, jsonLogic.apply("{\"in\": [2, {\"var\": \"list\"}]}", data));
-  }
+    @Test
+    public void testInVariableInt() throws JsonLogicException {
+        Map data = Collections.singletonMap("list", Arrays.asList(1, 2, 3));
+        assertEquals(true, jsonLogic.apply("{\"in\": [2, {\"var\": \"list\"}]}", data));
+    }
 
-  @Test
-  public void testNotInVariableInt() throws JsonLogicException {
-    Map data = Collections.singletonMap("list", Arrays.asList(1, 2, 3));
-    assertEquals(false, jsonLogic.apply("{\"in\": [4, {\"var\": \"list\"}]}", data));
-  }
+    @Test
+    public void testNotInVariableInt() throws JsonLogicException {
+        Map data = Collections.singletonMap("list", Arrays.asList(1, 2, 3));
+        assertEquals(false, jsonLogic.apply("{\"in\": [4, {\"var\": \"list\"}]}", data));
+    }
 
-  @Test
-  public void testAllVariables() throws JsonLogicException {
-    Map data = Stream.of(new Object[][] {
-      new Object[] {"list", Arrays.asList(1, 2, 3)},
-      new Object[] {"value", 3}
-    }).collect(Collectors.toMap(o -> o[0], o -> o[1]));
+    @Test
+    public void testAllVariables() throws JsonLogicException {
+        Map data = Stream.of(new Object[][]{
+                new Object[]{"list", Arrays.asList(1, 2, 3)},
+                new Object[]{"value", 3}
+        }).collect(Collectors.toMap(o -> o[0], o -> o[1]));
 
-    assertEquals(true, jsonLogic.apply("{\"in\": [{\"var\": \"value\"}, {\"var\": \"list\"}]}", data));
-  }
+        assertEquals(true, jsonLogic.apply("{\"in\": [{\"var\": \"value\"}, {\"var\": \"list\"}]}", data));
+    }
 
-  @Test
-  public void testSingleArgument() throws JsonLogicException {
-    assertFalse((boolean) jsonLogic.apply("{\"in\": [\"Spring\"]}", null));
-  }
+    @Test
+    public void testSingleArgument() throws JsonLogicException {
+        assertFalse((boolean) jsonLogic.apply("{\"in\": [\"Spring\"]}", null));
+    }
 
-  @Test
-  public void testBadSecondArgument() throws JsonLogicException {
-    assertFalse((boolean) jsonLogic.apply("{\"in\": [\"Spring\", 3]}", null));
-  }
+    @Test
+    public void testBadSecondArgument() throws JsonLogicException {
+        assertFalse((boolean) jsonLogic.apply("{\"in\": [\"Spring\", 3]}", null));
+    }
 }
