@@ -1,8 +1,8 @@
 package io.github.jamsesso.jsonlogic;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class MathExpressionTests {
   private static final JsonLogic jsonLogic = new JsonLogic();
